store event data
