+++ agent/native_windows/inc/com/centreon/agent/windows_util.hh
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2024 Centreon
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * For more information : contact@centreon.com
+ */
+
+#ifndef CENTREON_AGENT_WINDOWS_UTIL_HH
+#define CENTREON_AGENT_WINDOWS_UTIL_HH
+
+namespace com::centreon::agent {
+std::string get_last_error_as_string();
+
+std::string lpwcstr_to_acp(LPCWSTR lpwstr);
+
+}  // namespace com::centreon::agent
+
+#endif
