+++ agent/test/CMakeLists.txt
@@ -32,7 +32,8 @@ if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
   set(SRC ${SRC_COMMON} config_test.cc check_linux_cpu_test.cc)
 else()
   set(SRC ${SRC_COMMON}
-    check_event_log_test.cc
+    check_event_log_data_test.cc
+    check_event_log_filter_test.cc
     check_windows_cpu_test.cc  
     check_windows_memory_test.cc 
     check_uptime_test.cc
